requirements.txt
tensorflow==2.15.0.post1
pandas==2.1.4
dvc==3.35.0
mlflow==2.9.2
notebook==7.0.6
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
python-box==7.1.1
PyYAML==6.0.1
tqdm==4.66.1
ensure==1.0.4
joblib==1.3.2
types-PyYAML==6.0.12.12
scipy==1.11.4
Flask==3.0.0
Flask-Cors==4.0.0
gdown==4.7.1